# zhiguang_new/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/profile/avatar":{"post":{"tags":["profile-controller"],"operationId":"uploadAvatar","requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseProfileResponse"}}}}}}},"/api/v1/knowposts/{id}/publish":{"post":{"tags":["know-post-controller"],"operationId":"publish","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}}}}},"/api/v1/knowposts/{id}/content/confirm":{"post":{"tags":["know-post-controller"],"operationId":"confirmContent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowPostContentConfirmRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}}}}},"/api/v1/knowposts/drafts":{"post":{"tags":["know-post-controller"],"operationId":"createDraft","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseKnowPostDraftCreateResponse"}}}}}}},"/api/v1/auth/token/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenRefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseTokenResponse"}}}}}}},"/api/v1/auth/send-code":{"post":{"tags":["auth-controller"],"operationId":"sendCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendCodeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseSendCodeResponse"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseAuthResponse"}}}}}}},"/api/v1/auth/password/reset":{"post":{"tags":["auth-controller"],"operationId":"resetPassword","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/PasswordResetRequest"}}],"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogoutRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/v1/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseAuthResponse"}}}}}}},"/api/v1/profile":{"patch":{"tags":["profile-controller"],"operationId":"patch","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfilePatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseProfileResponse"}}}}}}},"/api/v1/knowposts/{id}":{"delete":{"tags":["know-post-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}}}},"patch":{"tags":["know-post-controller"],"operationId":"patchMetadata","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowPostPatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponse"}}}}}}},"/api/v1/knowposts/{id}/visibility":{"patch":{"tags":["know-post-controller"],"operationId":"patchVisibility","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowPostVisibilityPatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}}}}},"/api/v1/knowposts/{id}/top":{"patch":{"tags":["know-post-controller"],"operationId":"patchTop","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowPostTopPatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseVoid"}}}}}}},"/api/v1/knowposts/mine":{"get":{"tags":["know-post-controller"],"operationId":"mine","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseFeedPageResponse"}}}}}}},"/api/v1/knowposts/feed":{"get":{"tags":["know-post-controller"],"operationId":"feed","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseFeedPageResponse"}}}}}}},"/api/v1/knowposts/detail/{id}":{"get":{"tags":["know-post-controller"],"operationId":"detail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseKnowPostDetailResponse"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth-controller"],"operationId":"me","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BaseResponseAuthUserResponse"}}}}}}}},"components":{"schemas":{"BaseResponseProfileResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/ProfileResponse"},"message":{"type":"string"}}},"ProfileResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatar":{"type":"string"},"bio":{"type":"string"},"zgId":{"type":"string"},"gender":{"type":"string"},"birthday":{"type":"string","format":"date"},"school":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"tagJson":{"type":"string"}}},"BaseResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"message":{"type":"string"}}},"KnowPostContentConfirmRequest":{"required":["etag","objectKey","sha256","size"],"type":"object","properties":{"objectKey":{"type":"string"},"etag":{"type":"string"},"size":{"type":"integer","format":"int64"},"sha256":{"type":"string"}}},"BaseResponseKnowPostDraftCreateResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/KnowPostDraftCreateResponse"},"message":{"type":"string"}}},"KnowPostDraftCreateResponse":{"type":"object","properties":{"id":{"type":"string"}}},"TokenRefreshRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"BaseResponseTokenResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/TokenResponse"},"message":{"type":"string"}}},"TokenResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"accessTokenExpiresAt":{"type":"string","format":"date-time"},"refreshToken":{"type":"string"},"refreshTokenExpiresAt":{"type":"string","format":"date-time"}}},"SendCodeRequest":{"required":["identifier","identifierType","scene"],"type":"object","properties":{"scene":{"type":"string","enum":["REGISTER","LOGIN","RESET_PASSWORD"]},"identifierType":{"type":"string","enum":["PHONE","EMAIL"]},"identifier":{"type":"string"}}},"BaseResponseSendCodeResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/SendCodeResponse"},"message":{"type":"string"}}},"SendCodeResponse":{"type":"object","properties":{"identifier":{"type":"string"},"scene":{"type":"string","enum":["REGISTER","LOGIN","RESET_PASSWORD"]},"expireSeconds":{"type":"integer","format":"int32"}}},"RegisterRequest":{"required":["code","identifier","identifierType"],"type":"object","properties":{"identifierType":{"type":"string","enum":["PHONE","EMAIL"]},"identifier":{"type":"string"},"code":{"type":"string"},"password":{"type":"string"},"agreeTerms":{"type":"boolean"}}},"AuthResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/AuthUserResponse"},"token":{"$ref":"#/components/schemas/TokenResponse"}}},"AuthUserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatar":{"type":"string"},"phone":{"type":"string"},"zhId":{"type":"string"},"birthday":{"type":"string","format":"date"},"school":{"type":"string"},"bio":{"type":"string"},"gender":{"type":"string"},"tagJson":{"type":"string"}}},"BaseResponseAuthResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/AuthResponse"},"message":{"type":"string"}}},"PasswordResetRequest":{"required":["code","identifier","identifierType","newPassword"],"type":"object","properties":{"identifierType":{"type":"string","enum":["PHONE","EMAIL"]},"identifier":{"type":"string"},"code":{"type":"string"},"newPassword":{"type":"string"}}},"LogoutRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"LoginRequest":{"required":["identifier","identifierType"],"type":"object","properties":{"identifierType":{"type":"string","enum":["PHONE","EMAIL"]},"identifier":{"type":"string"},"code":{"type":"string"},"password":{"type":"string"}}},"ProfilePatchRequest":{"type":"object","properties":{"nickname":{"maxLength":64,"minLength":1,"type":"string"},"bio":{"maxLength":512,"minLength":0,"type":"string"},"gender":{"pattern":"(?i)MALE|FEMALE|OTHER|UNKNOWN","type":"string"},"birthday":{"type":"string","format":"date"},"zgId":{"pattern":"^[a-zA-Z0-9_]{4,32}$","type":"string"},"school":{"maxLength":128,"minLength":0,"type":"string"},"tagJson":{"type":"string"}}},"KnowPostPatchRequest":{"type":"object","properties":{"title":{"type":"string"},"tagId":{"type":"integer","format":"int64"},"tags":{"maxItems":20,"minItems":0,"type":"array","items":{"type":"string"}},"imgUrls":{"maxItems":20,"minItems":0,"type":"array","items":{"type":"string"}},"visible":{"type":"string"},"isTop":{"type":"boolean"},"description":{"type":"string"}}},"KnowPostVisibilityPatchRequest":{"required":["visible"],"type":"object","properties":{"visible":{"type":"string"}}},"BaseResponseVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"type":"object"},"message":{"type":"string"}}},"KnowPostTopPatchRequest":{"required":["isTop"],"type":"object","properties":{"isTop":{"type":"boolean"}}},"BaseResponseFeedPageResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/FeedPageResponse"},"message":{"type":"string"}}},"FeedItemResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"coverImage":{"type":"string"},"tags":{"type":"array","items":{"type":"string"}},"authorAvatar":{"type":"string"},"authorNickname":{"type":"string"},"tagJson":{"type":"string"},"likeCount":{"type":"integer","format":"int64"},"favoriteCount":{"type":"integer","format":"int64"},"liked":{"type":"boolean"},"faved":{"type":"boolean"},"isTop":{"type":"boolean"}}},"FeedPageResponse":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/FeedItemResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"hasMore":{"type":"boolean"}}},"BaseResponseKnowPostDetailResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/KnowPostDetailResponse"},"message":{"type":"string"}}},"KnowPostDetailResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"contentUrl":{"type":"string"},"images":{"type":"array","items":{"type":"string"}},"tags":{"type":"array","items":{"type":"string"}},"authorId":{"type":"string"},"authorAvatar":{"type":"string"},"authorNickname":{"type":"string"},"authorTagJson":{"type":"string"},"likeCount":{"type":"integer","format":"int64"},"favoriteCount":{"type":"integer","format":"int64"},"liked":{"type":"boolean"},"faved":{"type":"boolean"},"isTop":{"type":"boolean"},"visible":{"type":"string"},"type":{"type":"string"},"publishTime":{"type":"string","format":"date-time"}}},"BaseResponseAuthUserResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"data":{"$ref":"#/components/schemas/AuthUserResponse"},"message":{"type":"string"}}}}},"x-openapi":{"x-setting":null,"x-markdownFiles":[]}}